safari-600.7-branch/LayoutTests/media/video-seek-multiple-expected.txt


Test that multiple seeks issued at the same time will result in only a single 'seeked' event.

EVENT(canplaythrough)
RUN(video.currentTime = 0.5)
RUN(video.currentTime = 1.5)
RUN(video.currentTime = 2.0)
EXPECTED (video.seeking == 'true') OK

EVENT(seeked)

EXPECTED (video.seeking == 'false') OK
EXPECTED (video.currentTime == '2') OK
END OF TEST
